Adds FakeNetworkPipeTest to rtc_unittests.
Also updating the tests so they pass. The previous target was not part
of the targets that runs on the try bots, so they started breaking
during a recent refactor. By adding it to the common target we'll avoid
regression like this in the future.

Bug: None
Change-Id: I55391d454bd52964875818be39cdaa853360ae81
Reviewed-on: https://webrtc-review.googlesource.com/c/123199
Commit-Queue: Sebastian Jansson <srte@webrtc.org>
Reviewed-by: Karl Wiberg <kwiberg@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#26820}