'Fixing' a few TODOs by removing them.
The one with enum.IntFlag is not feasible. An attempt is done here:
https://webrtc-review.googlesource.com/c/src/+/133884

It requires re-writing QualityAssessment to Python3 which is too much
work for little benefit. (I tried, but couldn't get the unit-tests to
pass for both 2 and 3.)

The second one is not a real todo.

TBR=alessiob@webrtc.org
NOPRESUBMIT=True
Bug=None
NOTRY=True

Change-Id: Ia25817533cd504c30490f86e4058f0b2d59dd39c
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/133908
Commit-Queue: Alex Loiko <aleloi@webrtc.org>
Reviewed-by: Alex Loiko <aleloi@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#27715}