Fix CE in rtp_generator. Add it to default build target
This is a temporary solution, as there are several other executables and
some tests in rtc_tools/BUILD.gn. Including all of them to default target
is not decided yet.

But as rtp_generator tends to be broken reguraly, It should be included
there at least for now.


Bug: webrtc:10807
Change-Id: I3acf5a93c74bf1e2474c6aaee35653efbb43d3a0
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/146080
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Ilya Nikolaevskiy <ilnik@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#28595}