Enable field trial LegacySimulcastLayerLimit by default
When using simulcast, if the source is too small, it might end up generating
layer sizes that are problematic for hardware encoders.

We can temporarily restore the old behavior that adapts the layer count to the source size until we fix the HW encoder behavior.
to fix HW encoder issues

Bug: webrtc:10849, chromium:990823
Change-Id: Ie1486c9209b408c797c92d1b319d4116fe77171b
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/148069
Commit-Queue: Florent Castelli <orphis@webrtc.org>
Reviewed-by: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#28761}