Add a test which breaks if libwebrtc.a don't pull in the right symbols.
It's impossible to ensure we are pulling in everything people
reasonably believe is used, but it should be a good chunk of it.

I don't plan to actually run this test on the bots, it's enough if
it is built (which it should, because I add it to the default set
of things to build).

Bug: webrtc:11027
Change-Id: I186936eeb450d2f63b3a5bed13189e84d5b3fb76
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/157175
Commit-Queue: Patrik Höglund <phoglund@webrtc.org>
Reviewed-by: Niels Moller <nisse@webrtc.org>
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#29518}