Move IceFieldTrials into own .h-file
Several patches for webrtc:10647 has split the
P2PTransportChannel class/file. This has had the
side effect of it being hard to share the IceFieldTrials-struct.

This patch moves that struct into own file so that can be included
from other components. This patch is a behavioral NOP.

BUG=webrtc:10647

Change-Id: If49cd4d919684a48dde3188a26baf20e4ff2cd8e
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/160301
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Jonas Oreland <jonaso@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#29876}