Suppress C5041 constexpr warning for MSVC 2019
Disable the C5041 warning which makes the build fail. This is a
C++17-only change and WebRTC doesn't support C++17 yet, so the code is
technically correct, but fails to build on MSVC 2019 and
warning-as-error active.

Also fix another warning-as-error build error with MSVC 2019 due to
ignoring the result of a [[nodiscard]] function.

No-Presubmit: True
Bug: webrtc:11275,webrtc:11276
Change-Id: I891a894ee87252f96e84fd8d282576f46907256f
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/165781
Commit-Queue: Mirko Bonadei <mbonadei@webrtc.org>
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Reviewed-by: Karl Wiberg <kwiberg@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#30244}