Add interface_id to rtc::Network
This patch adds an interface_id property
to rtc::Network. It is an enumeration of the
interface names that are present.

This enables a local ICE agent to keep track
of which connections are using which interfaces,
something that is useful for predicting how
connections behave.

This is part 1 of https://webrtc-review.googlesource.com/c/src/+/85520

BUG: webrtc:9446
Change-Id: Ia6ec1f14ac240799fb1be49d67d82e2733e87acf
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/171061
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Jonas Oreland <jonaso@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#30882}