Add in missing protobuf code.
Turns out the import of histogram_pb2 failed not on the stub itself
(which I thought for a long, long time), but because of the protobuf
support code it includes in turn. This is a drawback of catching
the ImportError in histogram_proto.py.

This has a decent chance of fixing the problem.

Tbr: mbonadei@webrtc.org
Bug: chromium:1029452
Change-Id: I3c07a362dcfd174a388b3cc34449c08951cea626
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/171860
Reviewed-by: Patrik Höglund <phoglund@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#30912}