Remove assoc_send_channel_lock_ from ChannelReceive.
Associating a send channel is done on the same thread as network packets
are routed, which (currently) is also where stats are reported from,
so we can get rid of the lock and just make sure that the class is used
correctly.

Moving forward, this thread will become the network thread, so we'll
need to take a closer look at options for delivering the stats without
adding contention.

Bug: webrtc:11993
Change-Id: Ia87e67e8ae90b1651ef4a69243cf05093a620ed4
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/212612
Commit-Queue: Tommi <tommi@webrtc.org>
Reviewed-by: Niels Moller <nisse@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#33618}